feat: move activity-ticket method to activity-ticket folder
