+++ src/activity/activity.module.ts
@@ -5,11 +5,12 @@ import { ActivityService } from './activity.service';
 import { ActivityInfrastructure } from './activity.infra';
 import { AuthModule } from '~/auth/auth.module';
 import { DefinitionModule } from '~/definition/definition.module';
+import { ActivityTicketInfrastructure } from './activity-ticket/activity-ticket.infra';
 
 @Module({
-  providers: [Logger, ActivityTicketService, ActivityService, ActivityInfrastructure],
+  providers: [Logger, ActivityTicketService, ActivityService, ActivityTicketService , ActivityInfrastructure, ActivityTicketInfrastructure],
   imports: [DefinitionModule, forwardRef(() => AuthModule)],
   controllers: [ActivityController],
-  exports: [ActivityService, ActivityModule]
+  exports: [ActivityService, ActivityTicketService , ActivityModule]
 })
 export class ActivityModule {}
