+++ src/activity/activity.service.ts
@@ -1,11 +1,10 @@
 import { Injectable } from '@nestjs/common';
-import { ActivityCollectionDTO, ActivityDto, FetchActivitiesResponseDto } from './dto/activity.dto';
 import { ActivityInfrastructure } from './activity.infra';
 import { InjectEntityManager } from '@nestjs/typeorm';
 import { EntityManager } from 'typeorm';
 import { Activity } from './entity/Activity';
 import { ActivitySessionTicketEnrollmentCount } from './view_entity/ActivitySessionTicketEnrollmentCount';
-import { FetchMemberRightActivityTicketDTO, FetchMemberRightActivityTicketQuery, MemberRightActivityTicketDataDto } from './dto/member-right-activity-ticket.dto';
+import { ActivityCollectionDTO, ActivityDto, FetchActivitiesResponseDto } from './dto/activity.dto';
 
 @Injectable()
 export class ActivityService {
@@ -39,11 +38,6 @@ export class ActivityService {
     res.totalCount = totalCount;
 
     return res;
-  }
-
-  async memberRightActivityTicket(dto: FetchMemberRightActivityTicketQuery): Promise<MemberRightActivityTicketDataDto> {
-    return this.activityInfra.getActivityTicketInfoByIdAndMemberId(this.entityManager,  dto.activityTicketId, dto.memberId, dto.sessionId)
-
   }
 
   private async getActivities(dto: ActivityCollectionDTO): Promise<[Activity[], number]> {
