+++ src/equity/dto/equity-activity-ticket.dto.ts
